photo/requirements.txt
Flask
rembg
opencv-python
Pillow
numpy


